+++ library/multichoicerecyclerview/src/main/java/com/davidecirillo/multichoicerecyclerview/MultiChoiceToolbar.java
@@ -2,6 +2,7 @@ package com.davidecirillo.multichoicerecyclerview;
 
 import android.content.res.TypedArray;
 import android.support.annotation.PluralsRes;
+import android.support.annotation.StringRes;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.util.TypedValue;
@@ -105,12 +106,27 @@ public class MultiChoiceToolbar {
         }
 
         /**
-         * Set the titles for the differente state of the toolbar default/multiChoice
+         * Set the titles for the different states of the toolbar default/multiChoice
          *
          * @param defaultTitle          Title when the toolbar shown is the default one
          * @param selectedQuantityTitle Title shown when some item are selected. Will use the plural-feature to let you define the correct format.
          * @return Builder so you can chain together setters and build
          */
+        public Builder setTitles(@StringRes int defaultTitle, @PluralsRes int selectedQuantityTitle) {
+            this.mDefaultToolbarTitle = mAppCompatActivity.getString(defaultTitle);
+            this.mSelectedQuantityTitle = selectedQuantityTitle;
+            return this;
+        }
+
+        /**
+         * Set the titles for the different states of the toolbar default/multiChoice
+         *
+         * @param defaultTitle          Title when the toolbar shown is the default one
+         * @param selectedQuantityTitle Title shown when some item are selected. Will use the plural-feature to let you define the correct format.
+         * @return Builder so you can chain together setters and build
+         * @deprecated use {@link #setTitles(int, int)} instead
+         */
+        @Deprecated
         public Builder setTitles(String defaultTitle, @PluralsRes int selectedQuantityTitle) {
             this.mDefaultToolbarTitle = defaultTitle;
             this.mSelectedQuantityTitle = selectedQuantityTitle;