+++ sample/src/main/java/com/davidecirillo/multichoicesample/sampleCustomView/SampleCustomActivity.java
@@ -6,19 +6,14 @@ import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
 
+import com.davidecirillo.multichoicerecyclerview.MultiChoiceAdapter;
 import com.davidecirillo.multichoicerecyclerview.MultiChoiceToolbar;
 import com.davidecirillo.multichoicesample.BaseActivity;
 import com.davidecirillo.multichoicesample.R;
 
 import java.util.ArrayList;
-import java.util.List;
 
 import butterknife.BindView;
-import rx.Observable;
-import rx.Subscriber;
-import rx.android.schedulers.AndroidSchedulers;
-import rx.functions.Func1;
-import rx.schedulers.Schedulers;
 
 public class SampleCustomActivity extends BaseActivity {
 
@@ -33,36 +28,30 @@ public class SampleCustomActivity extends BaseActivity {
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+
         setUpMultiChoiceRecyclerView();
+    }
+
+    @Override
+    protected void onSaveInstanceState(Bundle outState) {
+        ((MultiChoiceAdapter) multiChoiceRecyclerView.getAdapter()).onSaveInstanceState(outState);
+        super.onSaveInstanceState(outState);
+    }
+
+    @Override
+    protected void onRestoreInstanceState(Bundle savedInstanceState) {
+        ((MultiChoiceAdapter) multiChoiceRecyclerView.getAdapter()).onRestoreInstanceState(savedInstanceState);
+        super.onRestoreInstanceState(savedInstanceState);
     }
 
     private void setUpMultiChoiceRecyclerView() {
 
         multiChoiceRecyclerView.setLayoutManager(new LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false));
 
-        getSampleMessageList()
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribeOn(Schedulers.newThread())
-                .subscribe(new Subscriber<List<MessageV0>>() {
-                    @Override
-                    public void onCompleted() {
-                    }
-
-                    @Override
-                    public void onError(Throwable e) {
-                    }
-
-                    @Override
-                    public void onNext(List<MessageV0> messageV0s) {
-                        setUpAdapter(messageV0s);
-                    }
-                });
-    }
-
-    private void setUpAdapter(List<MessageV0> messageV0s) {
         MultiChoiceToolbar multiChoiceToolbar =
                 new MultiChoiceToolbar.Builder(SampleCustomActivity.this, toolbar)
-                        .setTitles(toolbarTitle(), "")
+                        .setTitles(getString(toolbarTitle()
+                        ), "")
                         .setMultiChoiceColours(R.color.colorPrimaryMulti, R.color.colorPrimaryDarkMulti)
                         .setDefaultIcon(R.drawable.ic_arrow_back_white_24dp, new View.OnClickListener() {
                             @Override
@@ -73,33 +62,30 @@ public class SampleCustomActivity extends BaseActivity {
                         .build();
 
         SampleCustomViewAdapter adapter = new SampleCustomViewAdapter(
-                new ArrayList<>(messageV0s),
+                getSampleMessageList(),
                 SampleCustomActivity.this
         );
         adapter.setMultiChoiceToolbar(multiChoiceToolbar);
         multiChoiceRecyclerView.setAdapter(adapter);
+
     }
 
-    private Observable<List<MessageV0>> getSampleMessageList() {
-        return Observable
-                .range(0, 30)
-                .map(new Func1<Integer, MessageV0>() {
-                    @Override
-                    public MessageV0 call(Integer integer) {
-                        return new MessageV0("Title message number " + integer, "Lorem ipsum dolor " + integer + " sit amet, consectetur adipiscing" +
-                                " elit. Donec id mi pharetra, porta felis sed, aliquam urna. Curabitur porta dolor lobortis semper dictum. " +
-                                "Vestibulum posuere velit nisl, at porta lectus condimentum vel. Duis pharetra auctor tempor. Proin feugiat turpis " +
-                                "vel tincidunt molestie. Pellentesque tincidunt felis vitae leo pharetra aliquam. Donec sapien ante, feugiat at " +
-                                "eleifend vel, laoreet vitae neque. Donec eu erat et diam fermentum congue a sed libero. Vivamus dapibus nunc nec " +
-                                "posuere elementum. Nunc eget mi sed est finibus ultricies. Nam ut sapien feugiat neque tempus feugiat.");
-                    }
-                })
-                .toList();
+    private ArrayList<MessageV0> getSampleMessageList() {
+        ArrayList<MessageV0> sampleList = new ArrayList<>();
+        for (int i = 0; i < 100; i++) {
+            sampleList.add(new MessageV0("Title message number " + i, "Lorem ipsum dolor " + i + " sit amet, consectetur adipiscing" +
+                    " elit. Donec id mi pharetra, porta felis sed, aliquam urna. Curabitur porta dolor lobortis semper dictum. " +
+                    "Vestibulum posuere velit nisl, at porta lectus condimentum vel. Duis pharetra auctor tempor. Proin feugiat turpis " +
+                    "vel tincidunt molestie. Pellentesque tincidunt felis vitae leo pharetra aliquam. Donec sapien ante, feugiat at " +
+                    "eleifend vel, laoreet vitae neque. Donec eu erat et diam fermentum congue a sed libero. Vivamus dapibus nunc nec " +
+                    "posuere elementum. Nunc eget mi sed est finibus ultricies. Nam ut sapien feugiat neque tempus feugiat."));
+        }
+        return sampleList;
     }
 
     @Override
-    protected String toolbarTitle() {
-        return getString(R.string.custom_selection_view);
+    protected int toolbarTitle() {
+        return R.string.custom_selection_view;
     }
 
     @Override